% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_chifed.R
\name{available_chifed_varcodes}
\alias{available_chifed_varcodes}
\title{Get information about variables currently available in the extracted Chicago
database}
\usage{
available_chifed_varcodes(db_connector, min_date = NULL, max_date = NULL)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A \code{tibble} in the form of the codebook, but with an additional
column pairing a \code{VAR_CODE_ID} to each \code{VAR_CODE} found in the observations
table
}
\description{
\code{available_chifed_varcodes()} returns a subset of the codebook table found
in the database table \code{CHIFED.CODEBOOK} containing records only for those
variables currently available in \code{CHIFED.OBS_ALL} (as determined by a query
at the time of this function being called), as well as an additional column
indicating which value of \code{VAR_CODE_ID} in \code{CHIFED.OBS_ALL} is associated
with which \code{VAR_CODE} and description.
}
