% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{build_config_filename}
\alias{build_config_filename}
\title{Build the full path to a YAML configuration file}
\usage{
build_config_filename(
  subdir_name = "economicist",
  yaml_name = "callReports.yml"
)
}
\arguments{
\item{subdir_name}{(default \code{economicist}) A subdirectory within your local
user configuration directory}

\item{yaml_name}{(default \code{callReports.yml}) A name for this package's
configuration file}
}
\value{
A constructed file path
}
\description{
This package keeps your settings for the database, ZIP folders,
and logging folder in a persistent configuration file located in your user
configuration directory as given by \code{rappdirs::user_config_dir()}, in an
optional subdirectory specified by the parameter \code{subdir_name}. This
function builds and returns the full path.
}
