% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_schedule_filename}
\alias{build_schedule_filename}
\title{Rebuild an FFIEC schedule filename from its relevant parts}
\usage{
build_schedule_filename(rep_date, sch_code, part_num, part_of)
}
\arguments{
\item{rep_date}{A date string in ISO \code{YYYY-MM-DD} format}

\item{sch_code}{An FFIEC schedule code}

\item{part_num}{A part number}

\item{part_of}{The number of parts the schedule described by these parameters
is divided into.}
}
\value{
A character valued filename constructed from the given parameters.
}
\description{
From the reporting date, schedule code, part number and part code given as
parameters, construct a filename in the pattern of those found in the zipped
FFIEC schedule files.
}
\details{
The \code{SUMMARY} table in the FFIEC database contains five columns that indicate
the dates, schedule codes, and part numbers of FFIEC schedule files that have
already been extracted to the database. This function takes parameters
describing a hypothetical schedule file, in the format they would be seen in
the aforementioned \code{SUMMARY} table, and outputs a filename in the pattern
\verb{FFIEC CDR Call Schedule \{sch_code\} \{rep_date\}\{part_str\}.txt}, where
\code{part_str} is simply the \code{NULL} string when \code{part_of} is equal to \code{1}, as only
those schedule files that are in fact divided into parts have their part
numbers indicated in their filename.
}
