% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{chifed_dates_in_db}
\alias{chifed_dates_in_db}
\title{What dates have already been extracted to the Chicago Database?}
\usage{
chifed_dates_in_db(db_connector)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A character vector containing dates of already-extracted datasets
in the Chicago database, in ISO 'YYYY-MM-DD' format
}
\description{
What dates have already been extracted to the Chicago Database?
}
