% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{copy_example_script}
\alias{copy_example_script}
\title{Copy an example script to the location of your choice and open it.}
\usage{
copy_example_script(script = NULL, open = TRUE)
}
\arguments{
\item{script}{One of the scripts listed by \code{list_example_scripts()}}

\item{copy}{(default \code{TRUE}) Copy the file to a writable location? If \code{FALSE}
the copy of the script stored with the package will be opened. You should be
asked for a different location to save it to if you edit it.}
}
\description{
For user convenience, a few sample scripts are included with this package
which can be listed using \code{list_example_scripts()}. This function copies
the script selected by the user
}
