% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_dir_if_not_exists}
\alias{create_dir_if_not_exists}
\title{Create a directory if it doesn't exist}
\usage{
create_dir_if_not_exists(newdir_path)
}
\arguments{
\item{newdir_path}{A character-valued path to a folder to create}
}
\description{
\code{create_dir_if_not_exists()} checks to see if the specified directory exists
already. If it does, it returns \code{NULL}, but creates it otherwise.
}
