% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_user_cfg_if_not_exists}
\alias{create_user_cfg_if_not_exists}
\title{Create a configuration file for this package in the local user configuration
directory}
\usage{
create_user_cfg_if_not_exists(
  subdir_name = "economicist",
  yaml_name = "callReports.yml"
)
}
\arguments{
\item{subdir_name}{(default \code{economicist}) A subdirectory of your local
user configuration directory, in which to place the configuration file
given by \code{yaml_name}}

\item{yaml_name}{(default \code{callReports.yml}) A name for the configuration file}
}
\description{
\code{create_user_cfg_if_not_exists()} uses \code{get_cfg_filename()} to determine the
location of the user configuration file for this package. If no configuration
file is found at that location, the template provided with the package is
copied to it.
}
