% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_connectors.R
\name{db_connector_sqlite}
\alias{db_connector_sqlite}
\title{Create a \code{function} that opens an SQLite connection on command.}
\usage{
db_connector_sqlite(sqlite_db_path, overwrite = FALSE)
}
\arguments{
\item{sqlite_db_path}{The path to an SQLite database file. Must be in an
existing directory. The file will be created if it does not exist.}
}
\value{
A \code{function} that opens a connection to the database given in
\code{sqlite_db_path}. It should be passed to other functions without the \verb{()}
}
\description{
\code{db_connect_sqlite()} returns a \code{function} that allows an SQLite connection
to be opened whenever necessary by other functions in this package.
}
\examples{
db_connector_sqlite('./db/ffiec.sqlite')
}
