% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ffiec.R
\name{detect_ffiec_cross_references}
\alias{detect_ffiec_cross_references}
\title{Detect variables in the FFIEC database that exist in multiple tables}
\usage{
detect_ffiec_cross_references(db_connector)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A \code{tibble} telling you which tables a variable appearing in multiple
parts of the FFIEC reporting form can be found
}
\description{
Detect variables in the FFIEC database that exist in multiple tables
}
\examples{
The database connector only needs to be created once in any given script.
ffiec_db <- sqlite_connector('./db/ffiec.sqlite')
detect_ffiec_cross_references(ffiec_db)
}
