% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{extract_all_chifed_zips}
\alias{extract_all_chifed_zips}
\title{List ZIP files containing Call Report data from the Chicago Fed}
\usage{
extract_all_chifed_zips(db_connector, chifed_zip_path)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{chifed_zip_path}{A single character value containing an existing folder
containing the ZIP files}

\item{overwrite_sqlite}{(default \code{FALSE}) Should we overwrite an existing
SQLite DB? This is meaningful only for SQLite databases.}
}
\value{
A character vector containing the full paths of valid ZIP files
}
\description{
\code{list_chifed_zips()} expects the path of a folder containing the bulk data
files for 1976-2000 offered at the \href{https://www.chicagofed.org/banking/financial-institution-reports/commercial-bank-data-complete-1976-2000}{Chicago Fed's website}
}
\examples{
extract_all_chifed_zips('./zips-chifed')
}
