% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{extract_chifed_codebook}
\alias{extract_chifed_codebook}
\title{Extract the Chicago Fed's codebook data to a database}
\usage{
extract_chifed_codebook(db_connector, codebook_zip)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{codebook_zip}{An existing copy of \code{MDRM.zip}
page accompanying the main data.}
}
\description{
\code{extract_chifed_codebook()} extracts the Chicago Fed's call report data
dictionary provided at (their website)\link{https://www.federalreserve.gov/apps/mdrm/download_mdrm.htm}
to a database specified in the given database connector function.
}
\examples{
The database connector only needs to be created once in any given script.
chifed_db <- sqlite_connector('./db/chifed.sqlite')
extract_chifed_codebook(chifed_db, './MDRM.zip')
}
