% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{extract_chifed_mdrm}
\alias{extract_chifed_mdrm}
\title{Extract the Chicago Fed's codebook data to a database}
\usage{
extract_chifed_mdrm(codebook_zip)
}
\arguments{
\item{codebook_zip}{An existing copy of \code{MDRM.zip}
page accompanying the main data.}
}
\description{
\code{extract_chifed_mdrm()} extracts the Chicago Fed's call report data
dictionary provided at (their website)\link{https://www.federalreserve.gov/apps/mdrm/download_mdrm.htm}
to the SQLite file specified in \code{set_sqlite_file()}.
}
\examples{
extract_chifed_mdrm("./MDRM.zip")
}
