% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{extract_chifed_zip}
\alias{extract_chifed_zip}
\title{Extract the Chicago Fed's data observations to a database}
\usage{
extract_chifed_zip(db_connector, zip_file = NULL)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{zip_file}{An existing ZIP file containing a valid XPT data file}
}
\description{
\code{extract_chifed_observations()} reads a ZIP file containing a valid XPT file.
}
\details{
ZIP files containing the data are available at the \href{https://www.chicagofed.org/banking/financial-institution-reports/commercial-bank-data-complete-1976-2000}{Chicago Fed}.
They should be saved (but not extracted) to an empty folder.
}
\examples{
# The database connector only needs to be created once in any given script.
chifed_db <- sqlite_connector('./db/chifed.sqlite')
extract_chifed_zip(chifed_db, './zips-chifed/call0003-zip.zip')
}
