% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_ffiec_codebook}
\alias{extract_ffiec_codebook}
\title{Extract the FFIEC codebook information for a given schedule file}
\usage{
extract_ffiec_codebook(sch_unzipped)
}
\arguments{
\item{sch_unzipped}{The path to a valid FFIEC schedule data file on disk}
}
\description{
\code{extract_ffiec_codebook()} accepts a text file containing tab-separated-value
bulk data provided by the FFIEC, and writes the variable names, their
descriptions, and the table where they can be found to a given database.
}
\details{
Most applications will not use this function as it is called by
\code{extract_ffiec_schedule()} as part of extracting the codebook and observations
simultaneously.

ZIP files containing valid schedule files for 2001-present are offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
# The database connector only needs to be created once in any given script.
ffiec_db <- db_connector_sqlite('./db/ffiec.sqlite')
extract_ffiec_codebook(ffiec_db, 'FFIEC CDR Call Schedule RCCII 06302002.txt')
}
