% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_ffiec_datestr}
\alias{extract_ffiec_datestr}
\title{Extract the date from an FFIEC schedule filename in \code{YYYY-MM-DD} format}
\usage{
extract_ffiec_datestr(filename)
}
\arguments{
\item{filename}{The name of an FFIEC zip or schedule file, existent or not.
Really just any character value with an eight-digit string that can be
interpreted as a date in \code{MMDDYYYY} format.}
}
\value{
A character value in \code{YYYY-MM-DD} format containing the date of the
report corresponding to that schedule filename.
}
\description{
Extract the date from an FFIEC schedule filename in \code{YYYY-MM-DD} format
}
\examples{
extract_ffiec_datestr('FFIEC CDR Call Schedule RCCII 06302002.txt')
}
