% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_ffiec_descs}
\alias{extract_ffiec_descs}
\title{Extract the variable descriptions from an FFIEC schedule file}
\usage{
extract_ffiec_descs(sch_unzipped)
}
\arguments{
\item{sch_unzipped}{The path to an FFIEC schedule file on disk.}
}
\value{
A character vector containing the variable descriptions found in the
second line of the file.
}
\description{
Extract the variable descriptions from an FFIEC schedule file
}
\examples{
extract_ffiec_descs('FFIEC CDR Call Schedule RCCII 06302002.txt')
}
