% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_ffiec_names}
\alias{extract_ffiec_names}
\title{Extract the variable names from an FFIEC schedule file}
\usage{
extract_ffiec_names(sch_unzipped)
}
\arguments{
\item{sch_unzipped}{The path to an FFIEC schedule file on disk.}
}
\value{
A character vector containing the variable names found in the top line.
}
\description{
Extract the variable names from an FFIEC schedule file
}
\examples{
extract_ffiec_names('FFIEC CDR Call Schedule RCCII 06302002.txt')
}
