% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_ffiec_schedule}
\alias{extract_ffiec_schedule}
\title{Extract the observations for a given FFIEC schedule file to a database}
\usage{
extract_ffiec_schedule(sch_unzipped)
}
\arguments{
\item{sch_unzipped}{A valid FFIEC schedule data file extracted to disk#'}
}
\description{
\code{extract_ffiec_schedule()} accepts a text file containing tab-separated-value
bulk data provided by the FFIEC, and writes it to a given database.
}
\details{
If any warnings are triggered about parsing issues, they are absorbed by the
function and an attempt is made to repair the schedule file. A second attempt
is then made, and a visible warning issued it was not resolved. Any data
extracted will still be written to the database, but you may wish to
investigate their integrity. Tests existing data sets as of 2022 June 17
have been working without warnings.

Most applications will not use this function as it is called by
\code{process_ffiec_schedule()} as part of extracting the codebook and observations
simultaneously.

ZIP files containing valid schedule files for 2001-present are offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
# The database connector only needs to be created once in any given script.
ffiec_db <- db_connector_sqlite('./db/ffiec.sqlite')
extract_ffiec_schedule(ffiec_db, 'FFIEC CDR Call Schedule RCCII 06302002.txt')
}
