% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_ffiec_tsv}
\alias{extract_ffiec_tsv}
\title{Extract the FFIEC codebook and all observations for a single schedule}
\usage{
extract_ffiec_tsv(zf, sch)
}
\arguments{
\item{zf}{A ZIP file containing schedule files}

\item{sch}{A valid FFIEC schedule data file inside the ZIP file \code{zf}}
}
\description{
\code{extract_ffiec_tsv()} accepts a ZIP file, extracts one of its enclosed
tab-separated-value schedule files, and writes the extracted data to a given
database.
}
\details{
ZIP files containing the bulk data 2001-present offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
extract_ffiec_tsv(
  "./zips-ffiec/FFIEC FFIEC CDR Call Bulk All Schedules 03312018.zip",
  "FFIEC CDR Call Schedule RCCII 06302002.txt"
)
}
