% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_ffiec_zips}
\alias{extract_ffiec_zips}
\title{Extract an entire directory of FFIEC ZIP files to a database}
\usage{
extract_ffiec_zips(ffiec_zip_path = get_ffiec_zip_dir())
}
\arguments{
\item{ffiec_zip_path}{(default \code{get_ffiec_zip_dir()}) A single character
value containing an existing folder containing the ZIP files}
}
\description{
\code{extract_ffiec_zips()} examines all ZIP files in \code{ffiec_zip_path} and
attempts to extract each enclosed schedule file to the database set by the
user. The underlying extraction function skips schedule files that have
already been successfully extracted to the database.
}
