% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_part_codes}
\alias{extract_part_codes}
\title{Extract the schedule part number from an FFIEC schedule filename}
\usage{
extract_part_codes(sch)
}
\arguments{
\item{sch}{The name of an FFIEC schedule file, existent or not.}
}
\value{
A numeric vector with two named values: \code{part_num} and \code{part_of},
with both being \code{1} if no part is indicated in the filename.
}
\description{
If data for a schedule is issued in one part (true for most), then it will
be determined to be part 1 of 1.
}
\examples{
> extract_part_code('FFIEC CDR Call Schedule RCB 03312013(2 of 2).txt')
part_num  part_of 
       1        2 
}
