% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_schedule_code}
\alias{extract_schedule_code}
\title{Extract the alphabetical schedule code from an FFIEC schedule filename}
\usage{
extract_schedule_code(sch)
}
\arguments{
\item{sch}{The name of an FFIEC schedule file, existent or not.}
}
\value{
A character value containing the alphabetical schedule code
corresponding to the given schedule filename.
}
\description{
Extract the alphabetical schedule code from an FFIEC schedule filename
}
\examples{
> extract_schedule_code('FFIEC CDR Call Schedule RCCII 06302002.txt')
[1] "RCCII"
}
