% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{extract_values}
\alias{extract_values}
\title{Extract the values from one line of an FFIEC schedule file}
\usage{
extract_values(sch_line)
}
\arguments{
\item{sch_line}{A line from an FFIEC tab-separated schedule file}
}
\value{
A character vector of the values in that line
}
\description{
Extract the values from one line of an FFIEC schedule file
}
\examples{
extract_values('12311  2031  298310')
[1] "12311" "2031" "298310"
}
