% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ffiec.R
\name{fetch_ffiec_extracted_schedules}
\alias{fetch_ffiec_extracted_schedules}
\title{Quickly determine which periods are already in the FFIEC database}
\usage{
fetch_ffiec_extracted_schedules()
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A character vector of dates in ISO YYYY-MM-DD format
}
\description{
Quickly determine which periods are already in the FFIEC database
}
