% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_ffiec.R
\name{fetch_ffiec_observations}
\alias{fetch_ffiec_observations}
\title{Query the FFIEC database for variables in a given database table}
\usage{
fetch_ffiec_observations(tbl_name, ...)
}
\arguments{
\item{tbl_name}{A valid table name containing FFIEC observations.}

\item{...}{Quoted variable names, separated by commas}
}
\value{
A \code{tibble} containing bank IDRSSD, report date, and the values of the
requested variables in wide format.
}
\description{
\code{query_ffiec_db()} connects to a given database and searches the given
schedule table for the requested variables.
}
\examples{
fetch_ffiec_observations("FFIEC.OBS_RC", "RCON2170", "RCON2948", "RCON3210")
}
