% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_metadata.R
\name{fetch_varcodes}
\alias{fetch_varcodes}
\title{Retrieve variable names from an index table}
\usage{
fetch_varcodes(db_connector)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A character vector of variable codes
}
\description{
\code{fetch_varcodes()} retrieves a character vector containing all the variable
codes that have thus far been assigned database ID values with \code{write_varcodes()}
}
\details{
Given the numerous variables in the source data, this library pivots the data
into long form before writing to the database. This offers the benefit of
allowing each table in the database to have a predictable description and
prevent issues with exceeding the native maximum column support of many
database engines.

However, pivoting the data creates a text column with a copy of the variable
code being stored for every one of the millions of observations in the data
set, significantly expanding the storage requirement of the database, and
potentially making queries run noticeably slower on standard spinning-platter
hard drives due to the increased scanning time.

This function is part of an effort to alleviate that problem by assigning
each variable code an integer ID value so that the storage requirements are
significantly smaller.
}
