% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ffiec_date_str_to_qtr_id}
\alias{ffiec_date_str_to_qtr_id}
\title{Convert a reporting date into an integer quarter index}
\usage{
ffiec_date_str_to_qtr_id(date_str)
}
\arguments{
\item{date_str}{A \code{Date} or character value in ISO \code{YYYY-MM-DD} format}
}
\value{
The index of the quarter since the beginning of the year 2001.
}
\description{
Convert a reporting date into an integer quarter index
}
