% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ffiec_qtr_id_to_date_str}
\alias{ffiec_qtr_id_to_date_str}
\title{Convert a reporting date index back into a date}
\usage{
ffiec_qtr_id_to_date_str(qtr_int)
}
\arguments{
\item{qtr_int}{An \code{integer} representing the index of a quarter since the
beginning of year 2001}
}
\value{
A \code{Date} object
}
\description{
Convert a reporting date index back into a date
}
