% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ffiec_db.R
\name{ffiec_schedules_in_db}
\alias{ffiec_schedules_in_db}
\title{Quickly determine which periods are already in the FFIEC database}
\usage{
ffiec_schedules_in_db(db_connector)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A character vector of dates in ISO YYYY-MM-DD format
}
\description{
Quickly determine which periods are already in the FFIEC database
}
