% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill_na_with_previous}
\alias{fill_na_with_previous}
\title{Replace \code{NA} values with last-seen non-\code{NA} value}
\usage{
fill_na_with_previous(x)
}
\arguments{
\item{x}{A vector suspected to contain \code{NA} values}
}
\value{
A vector containing all non-\code{NA} values from \code{x}, but with all \code{NA}
values replaced with the "last-seen" non-\code{NA} value up to that point in the
vector
}
\description{
Given a vector \code{x} containing NA values, accumulate a new output vector by
progressively replacing each NA value with the value of the last non-NA
value of an earlier index
}
\examples{
fill_na_with_previous(c(1, 2, NA, NA, 3)) ## [1] 1 2 2 2 3
}
