% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{generate_ffiec_sch_name}
\alias{generate_ffiec_sch_name}
\title{Generate an FFIEC TSV schedule from a reporting date, schedule code, and
(for applicable schedules) part numbers.}
\usage{
generate_ffiec_sch_name(ymd_chr, sch_code, part_num = 1, part_of = 1)
}
\arguments{
\item{ymd_chr}{A character reporting date in \code{YYYY-MM-DD} format.}

\item{sch_code}{A character schedule code}

\item{part_num}{(default \code{1}) Number of the applicable schedule part}

\item{part_of}{(default \code{1}) An integer number of parts to the schedule}
}
\value{
A character-valued ZIP filename for the given date
}
\description{
Generate an FFIEC TSV schedule from a reporting date, schedule code, and
(for applicable schedules) part numbers.
}
\examples{
generate_sch_name("2016-03-31", "RC")
generate_sch_name("2016-03-31", "RCO")
}
