% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{generate_ffiec_zip_name}
\alias{generate_ffiec_zip_name}
\title{Generate the path to an FFIEC ZIP filename from a reporting date}
\usage{
generate_ffiec_zip_name(ymd_chr, full_path = TRUE)
}
\arguments{
\item{ymd_chr}{A character reporting date in \code{YYYY-MM-DD} format.}

\item{full_path}{(default \code{TRUE}) Return the full path to the ZIP file?}
}
\value{
The full path to the FFIEC ZIP filename for the given date
}
\description{
Generate the path to an FFIEC ZIP filename from a reporting date
}
\examples{
generate_zip_path("2016-03-31")
}
