% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{generate_log_name}
\alias{generate_log_name}
\title{Build a log filename with a prefix and the time it was initiated.}
\usage{
generate_log_name(prefix)
}
\arguments{
\item{prefix}{A character value to begin the name of the file. May contain
slashes if you want the generated log file to go in a different folder than
specified by \code{getwd()}}
}
\value{
A character value in the form \verb{\{prefix\}_\{datestr\}.log.txt}
}
\description{
Build a log filename with a prefix and the time it was initiated.
}
\examples{
generate_log_name("extraction_attempt") ## [1] "extraction_attempt_20220628_141233.log.txt"
}
