% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{get_all_chifed_varnames}
\alias{get_all_chifed_varnames}
\title{Get a data frame with all XPT files and variables for the Chicago Fed data}
\usage{
get_all_chifed_varnames(chifed_zip_path)
}
\arguments{
\item{chifed_zip_path}{Folder containing the Chicago Fed zip files}
}
\value{
A \code{tibble} with two columns, one with the name of the XPT file and
another with the names of the variables found in that XPT file.
}
\description{
Get a data frame with all XPT files and variables for the Chicago Fed data
}
