% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_cfg_filename}
\alias{get_cfg_filename}
\title{Build the full path to a YAML configuration file}
\usage{
get_cfg_filename(subdir_name = "economicist", yaml_name = "callReports.yml")
}
\arguments{
\item{subdir_name}{(default \code{economicist}) A subdirectory of your local
user configuration directory, in which to place the configuration file
given by \code{yaml_name}}

\item{yaml_name}{(default \code{callReports.yml}) A name for the configuration file}
}
\value{
A full constructed file path to the configuration file given by
\code{subdir_name} and \code{yaml_name}
}
\description{
This package keeps your settings for the database, ZIP folders,
and logging folder in a persistent configuration file located in your user
configuration directory as given by \code{rappdirs::user_config_dir()}, in an
optional subdirectory specified by the parameter \code{subdir_name}. This
function builds and returns the full path.
}
