% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{get_chifed_varnames}
\alias{get_chifed_varnames}
\title{Get a data from with variables for a single Chicago Fed dataset}
\usage{
get_chifed_varnames(zf)
}
\arguments{
\item{zf}{A zip file containing an \code{XPT} file from the Chicago Fed}
}
\value{
A \code{tibble} with the \code{XPT} filename in one column and all of its
variables names in the other
}
\description{
Get a data from with variables for a single Chicago Fed dataset
}
