% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_chifed_zip_dir}
\alias{get_chifed_zip_dir}
\title{Get the directory where the Chicago Fed ZIP files are stored}
\usage{
get_chifed_zip_dir()
}
\value{
A directory character value
}
\description{
Get the directory where the Chicago Fed ZIP files are stored
}
