% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_setting}
\alias{get_config_setting}
\title{Fetch selected configuration settings for the package}
\usage{
get_config_setting(key, cfg_file = build_config_filename())
}
\arguments{
\item{cfg_file}{The path to a YAML configuration file containing the settings
given in \code{...}}

\item{...}{The character-valued names of the configuration settings you wish
to retrieve}
}
\value{
A named list with the values of the requested configuration settings
}
\description{
Fetch selected configuration settings for the package
}
