% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_ffiec_zip_dir}
\alias{get_ffiec_zip_dir}
\title{Get the directory where the FFIEC ZIP files are stored}
\usage{
get_ffiec_zip_dir()
}
\value{
A directory character value
}
\description{
Get the directory where the FFIEC ZIP files are stored
}
