% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_logging_dir}
\alias{get_logging_dir}
\title{Get the directory where the extraction and query logs are stored}
\usage{
get_logging_dir()
}
\value{
A directory character value
}
\description{
Get the directory where the extraction and query logs are stored
}
