% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_sqlite_filename}
\alias{get_sqlite_filename}
\title{Get the name of the SQLite database}
\usage{
get_sqlite_filename()
}
\value{
A directory character value
}
\description{
Get the name of the SQLite database
}
