% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{list_all_chifed_zips}
\alias{list_all_chifed_zips}
\title{List ZIP files containing Call Report data from the Chicago Fed}
\usage{
list_all_chifed_zips(chifed_zip_path)
}
\arguments{
\item{chifed_zip_path}{A single character value containing an existing folder
containing the ZIP files}
}
\value{
A character vector containing the full paths of valid ZIP files
}
\description{
\code{list_chifed_zips()} expects the path of a folder containing the bulk data
files for 1976-2000 offered at the \href{https://www.chicagofed.org/banking/financial-institution-reports/commercial-bank-data-complete-1976-2000}{Chicago Fed's website}
}
\examples{
list_all_chifed_zips('./zips-chifed')
}
