% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{list_example_scripts}
\alias{list_example_scripts}
\title{List the available example scripts for this package}
\usage{
list_example_scripts()
}
\value{
A list of sample scripts available for opening with \code{open_example_script()}
}
\description{
For user convenience, a few sample scripts are included with this package
which can be opened using \code{open_example_script()}. This function lists the
available sample scripts.
}
