% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_listing.R
\name{list_ffiec_tsvs}
\alias{list_ffiec_tsvs}
\title{List schedule files within an FFIEC Call Report ZIP}
\usage{
list_ffiec_tsvs(zip_file)
}
\arguments{
\item{zip_file}{A ZIP file containing bulk data provided by the FFIEC}
}
\value{
A \code{tibble} with the name of  containing the schedule filenames contained in \code{zip_file}
}
\description{
\code{list_ffiec_tsvs()} lists the schedule files contained within a given
ZIP file provided by the FFIEC.
}
\details{
ZIP files containing the bulk data 2001-present offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
list_ffiec_tsvs('FFIEC CDR Call Bulk All Schedules 06302004.zip')
}
