% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_listing.R
\name{list_ffiec_zips}
\alias{list_ffiec_zips}
\title{List ZIP files containing Call Report data from the FFIEC}
\usage{
list_ffiec_zips(ffiec_zip_path)
}
\arguments{
\item{ffiec_zip_path}{A single character value containing an existing folder
containing the ZIP files}
}
\value{
A character vector containing the full paths of valid ZIP files
}
\description{
\code{list_ffiec_zips()} lists valid ZIP files for extraction in a given directory.
}
\details{
ZIP files containing the bulk data 2001-present offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
list_ffiec_zips('./zips-ffiec')
}
