% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_listing.R
\name{list_ffiec_zips_and_tsvs}
\alias{list_ffiec_zips_and_tsvs}
\title{List ZIP and schedule files containing FFIEC bulk call report data}
\usage{
list_ffiec_zips_and_tsvs(ffiec_zip_path)
}
\arguments{
\item{ffiec_zip_path}{A single character value containing an existing folder
containing the ZIP files}
}
\value{
A \code{tibble} containing the full paths of valid ZIP files in one column
paired with the schedule files contained therein in the other
}
\description{
\code{list_ffiec_zips_and_tsvs()} provides an easiy traversable data frame
full of zip-schedule pairs that can be used by \code{extract_ffiec_tsv()}
}
\details{
ZIP files containing the bulk data 2001-present offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
list_ffiec_zips_and_tsvs("./zips-ffiec")
}
