% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{parse_ffiec_schedule}
\alias{parse_ffiec_schedule}
\title{Parse the observations in a given FFIEC schedule file for processing in \code{R}}
\usage{
parse_ffiec_schedule(sch_unzipped)
}
\arguments{
\item{sch_unzipped}{A valid FFIEC schedule data file extracted to disk#'}

\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A \code{tibble} containing the schedule data in wide form
}
\description{
\code{parse_ffiec_schedule()} accepts a text file containing tab-separated-value
bulk data provided by the FFIEC, and parses them into an object that can
be manipulated in \code{R}.
}
\details{
Each valid line in the schedule ends in a blank tab-separated value that has
no name in the header. This triggers a message about giving that column a
name when it is parsed. That message is suppressed by explicitly providing a
method for assigning a name to that blank column. That column is deselected
before the data frame is returned.

This function is unlikely to ever be called by the user except for diagnostics.

ZIP files containing valid schedule files for 2001-present are offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
parse_ffiec_schedule('FFIEC CDR Call Schedule RCCII 06302002.txt')
}
