% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{parse_ffiec_tsv.by_params}
\alias{parse_ffiec_tsv.by_params}
\title{Parse the FFIEC TSV file given by relevant parameters rather than filenames}
\usage{
parse_ffiec_tsv.by_params(ymd_chr, sch_code, part_num = 1, part_code = 1)
}
\arguments{
\item{ymd_chr}{A reporting date}

\item{sch_code}{A schedule code}

\item{part_num}{(default \code{1}) A part number, if applicable}

\item{part_code}{(default \code{2}) The number of parts to the schedule, if applicable}
}
\description{
Parse the FFIEC TSV file given by relevant parameters rather than filenames
}
