% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_mdy_substring}
\alias{parse_mdy_substring}
\title{Find and parse a date substring}
\usage{
parse_mdy_substring(chr, format)
}
\arguments{
\item{A}{character scalar containing a substring that can be interpreted
as a date.}
}
\value{
A character value in \code{YYYY-MM-DD} format containing the date of the
report corresponding to that schedule filename.
}
\description{
Find and parse a date substring
}
\examples{
> parse_mdy_substring('FFIEC CDR Call Schedule RCCII 06302002.txt')

}
