% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_tsv_line}
\alias{parse_tsv_line}
\title{Extract the values from one line of an FFIEC schedule file}
\usage{
parse_tsv_line(tsv_line)
}
\arguments{
\item{sch_line}{A line from an FFIEC tab-separated schedule file}
}
\value{
A character vector of the values in that line
}
\description{
Extract the values from one line of an FFIEC schedule file
}
\examples{
parse_tsv_line("12311  2031  298310") ## [1] "12311" "2031" "298310"
}
