% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{process_ffiec_schedule}
\alias{process_ffiec_schedule}
\title{Extract the FFIEC codebook and all observations for a single schedule}
\usage{
process_ffiec_schedule(db_connector, zf, sch)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{zf}{A ZIP file containing schedule files}

\item{sch}{A valid FFIEC schedule data file inside the ZIP file \code{zf}}
}
\description{
\code{process_ffiec_schedule()} accepts a ZIP file, extracts one of its enclosed
tab-separated-value schedule files, and writes the extracted data to a given
database.
}
\details{
ZIP files containing the bulk data 2001-present offered at the
\href{https://cdr.ffiec.gov/public/PWS/DownloadBulkData.aspx}{FFIEC Bulk Data Download Service}
}
\examples{
The database connector only needs to be created once in any given script.
ffiec_db <- sqlite_connector('./db/ffiec.sqlite')
process_ffiec_schedule(
  ffiec_db,
  './zips-ffiec/FFIEC FFIEC CDR Call Bulk All Schedules 03312018.zip',
  'FFIEC CDR Call Schedule RCCII 06302002.txt'
)
}
