% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{qtr_end}
\alias{qtr_end}
\title{Get the last day of a calendar quarter}
\usage{
qtr_end(ymd_chr = Sys.Date())
}
\arguments{
\item{dt}{A \code{Date}, or a character value in ISO \code{YYYY-MM-DD} format}
}
\value{
The \code{Date} of the final day of the quarter \code{dt} is in.
}
\description{
Get the last day of a calendar quarter
}
