% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{qtrs_since_2000}
\alias{qtrs_since_2000}
\title{Convert a date to a simple integer representing quarters since Q1 2001.}
\usage{
qtrs_since_2000(dt_str)
}
\arguments{
\item{dt}{A date in ISO \code{YYYY-MM-DD} format}
}
\value{
A single integer value, where \code{2001-03-31} --> \code{1},
\code{2001-06-30} --> \code{2}, and so on...
}
\description{
Convert a date to a simple integer representing quarters since Q1 2001.
}
