% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_chifed_db.R
\name{query_chifed_db}
\alias{query_chifed_db}
\title{Retreive Chicago Fed observations from a database}
\usage{
query_chifed_db(db_connector, ...)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{...}{Quoted variable names, separated by commas.}
}
\value{
A \code{tibble()} containing bank ID, report date, and values for the
variables requested in \code{...}
}
\description{
Retreive Chicago Fed observations from a database
}
\examples{
# The database connector only needs to be created once in any given script.
chifed_db <- sqlite_connector('./db/chifed.sqlite')
query_chifed_db('RCON2170', 'RCON2950', 'RCON3210')
}
