% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ffiec_db.R
\name{query_ffiec_db}
\alias{query_ffiec_db}
\title{Query the FFIEC database for variables in a given database table}
\usage{
query_ffiec_db(db_connector, sch_code, ...)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{sch_code}{A valid alphabetical schedule code}

\item{...}{Quoted variable names, separated by commas}
}
\value{
A \code{tibble} containing bank ID, report date, and the values of the
requested variables.
}
\description{
\code{query_ffiec_db()} connects to a given database and searches the given
schedule table for the requested variables.
}
\examples{
# The database connector only needs to be created once in any given script.
ffiec_db <- sqlite_connector('./db/ffiec.sqlite')
query_ffiec_db(ffiec_db, 'RCON2170', 'RCON2948', 'RCON3210')
}
