% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ffiec_db.R
\name{query_multischedule_variable}
\alias{query_multischedule_variable}
\title{Query a variable from all FFIEC tables where it exists}
\usage{
query_multischedule_variable(db_connector, var_name)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{var_name}{A variable that exists in multiple schedules, as determined
by the results of \code{search_ffiec_codebook()}}
}
\value{
A \code{tibble} in the same format as results of \code{query_ffiec_db()}, but
the column names represent the schedule in which the column is from
}
\description{
In case you want to know which FFIEC schedule contains the most instances
of a given variable, you can use this function to query all the tables where
it exists. You can view the results to determine which schedule you should
query for that variable in the future.
}
\examples{
# The database connector only needs to be created once in any given script.
ffiec_db <- sqlite_connector('./db/ffiec.sqlite')
query_multischedule_variable(ffiec_db, 'RCON2170')
}
