% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{read_chifed_zip}
\alias{read_chifed_zip}
\title{Extract the observations from a Chicago Fed ZIPped XPT file.}
\usage{
read_chifed_zip(zip_file)
}
\arguments{
\item{zip_file}{the path to a ZIP file on disk containing a valid XPT file}
}
\value{
A \code{tibble} containing the observations in wide form.
}
\description{
ZIP files containing the data are available at the \href{https://www.chicagofed.org/banking/financial-institution-reports/commercial-bank-data-complete-1976-2000}{Chicago Fed}.
They should be saved (but not extracted) to an empty folder.
}
