% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{read_ffiec_varcodes}
\alias{read_ffiec_varcodes}
\title{Extract the variable names from an FFIEC schedule file}
\usage{
read_ffiec_varcodes(sch_unzipped)
}
\arguments{
\item{sch_unzipped}{The path to an FFIEC schedule file on disk.}
}
\value{
A character vector containing the variable names found in the top line.
}
\description{
Extract the variable names from an FFIEC schedule file
}
\examples{
read_ffiec_varcodes('FFIEC CDR Call Schedule RCCII 06302002.txt')
}
