% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{read_ffiec_vardescs}
\alias{read_ffiec_vardescs}
\title{Extract the variable descriptions from an FFIEC schedule file}
\usage{
read_ffiec_vardescs(sch_unzipped)
}
\arguments{
\item{sch_unzipped}{The path to an FFIEC schedule file on disk.}
}
\value{
A character vector containing the variable descriptions found in the
second line of the file.
}
\description{
Extract the variable descriptions from an FFIEC schedule file
}
\examples{
read_ffiec_vardescs('FFIEC CDR Call Schedule RCCII 06302002.txt')
}
