% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{repair_colnames}
\alias{repair_colnames}
\title{Repair any invalid column names}
\usage{
repair_colnames(nms, prefix = "UNNAMED")
}
\arguments{
\item{nm}{A character-valued variable name}

\item{idx}{The index of that variable in the data set}
}
\value{
A valid and unique variable name
}
\description{
\code{vroom::vroom()} display a message about renaming variables with invalid
names if no rule is provided for repairing them. This function provides such
a rule, and can be used as the \code{name_repair} parameter of those functions.
}
