% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{reset_user_cfg}
\alias{reset_user_cfg}
\title{Delete all user configurations for this package}
\usage{
reset_user_cfg(subdir_name = "economicist", yaml_name = "callReports.yml")
}
\arguments{
\item{subdir_name}{(default \code{economicist}) A subdirectory of your local
user configuration directory, in which to place the configuration file
given by \code{yaml_name}}

\item{yaml_name}{(default \code{callReports.yml}) A name for the configuration file}
}
\description{
Delete all user configurations for this package
}
