% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{schedule_name_components}
\alias{schedule_name_components}
\title{Get the components of a schedule filename in easily-referenced format.}
\usage{
schedule_name_components(sch)
}
\arguments{
\item{sch}{The name of an FFIEC tab-separated schedule file.}
}
\value{
A named vector containing a value for \code{report_date}, \code{sch_code},
\code{part_num} (default \code{1}) and \code{part_of} (default \code{1})
and number of parts available for the schedule.
}
\description{
Get the components of a schedule filename in easily-referenced format.
}
\examples{
> schedule_name_components('FFIEC CDR Call Schedule RCB 03312012(1 of 2).txt')
 report_date     sch_code     part_num      part_of 
"2012-03-31"        "RCB"          "1"          "2" 
}
