% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_chifed_db.R
\name{search_chifed_codebook}
\alias{search_chifed_codebook}
\title{Search the Chicago Fed data codebook in the database, by name or description}
\usage{
search_chifed_codebook(db_connector, var_name = NULL, var_desc = NULL)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{var_name}{Either a four-letter variable prefix or full 8-character name}

\item{var_desc}{Any substring (case-insensitive) you hope to find in a
variable description}
}
\value{
A \code{tibble} containing matching results in the codebook
}
\description{
Either a \code{var_name} or \code{var_desc} must be provided.
}
\examples{
The database connector only needs to be created once in any given script.
chifed_db <- sqlite_connector('./db/chifed.sqlite')
search_chifed_codebook(var_name = 'RCON')
search_chifed_codebook(var_name = 'RCON2950')
search_chifed_codebook(var_desc = 'TOTAL LIABILITIES')
}
