% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_chifed_zip_dir}
\alias{set_chifed_zip_dir}
\title{Set the directory where the Chicago Fed ZIP files are stored}
\usage{
set_chifed_zip_dir(path = NULL)
}
\value{
A directory character value
}
\description{
Set the directory where the Chicago Fed ZIP files are stored
}
