% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_ffiec_zip_dir}
\alias{set_ffiec_zip_dir}
\title{Set the directory where the FFIEC ZIP files are stored}
\usage{
set_ffiec_zip_dir(path = NULL)
}
\value{
A directory character value
}
\description{
Set the directory where the FFIEC ZIP files are stored
}
