% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_logging_dir}
\alias{set_logging_dir}
\title{Get the directory where the extraction and query logs are stored}
\usage{
set_logging_dir(path = NULL)
}
\value{
A directory character value
}
\description{
Get the directory where the extraction and query logs are stored
}
