% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_sqlite_filename}
\alias{set_sqlite_filename}
\title{Get the name of the SQLite database}
\usage{
set_sqlite_filename(path = NULL)
}
\value{
A directory character value
}
\description{
Get the name of the SQLite database
}
