% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{split_ffiec_sch_name}
\alias{split_ffiec_sch_name}
\title{Split an FFIEC TSV schedule filename into its relevant constituent parts}
\usage{
split_ffiec_sch_name(tsv_filename)
}
\arguments{
\item{tsv_filename}{A filename matching the format of TSV files found within
FFIEC ZIPs.}
}
\value{
A named \code{list} with, at minimum \code{ymd_chr} and \code{sch_code} representing the reporting
date and schedule code, respectively, and (if applicable) \code{part_num} and \code{part_of}.
}
\description{
This function takes a file in the \code{glue} format: \verb{FFIEC CDR Call Schedule \{sch_code\} \{mm\}\{dd\}\{yyyy\}\{parts_chr\}.txt},
where \code{parts_chr} = \verb{(\{part_num\} of \{part_of\})} if applicable, and returns
a named list containing the report date in ISO \code{YYYY-MM-DD} format, the
schedule code, and (if applicable) the part number and number of parts.
}
\examples{
sch_filename <- generate_ffiec_sch_name("2012-06-30", "RCO", 1, 2)
split_ffiec_sch_name(sch_filename)
}
