% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_is_blank}
\alias{str_is_blank}
\title{Is a character value \code{NULL}, of length zero, or just whitespace?}
\usage{
str_is_blank(val)
}
\arguments{
\item{val}{A scalar character value}
}
\value{
\code{TRUE} if \code{val} is either \code{NULL}, a string of length zero, or a string
containing only whitespace
}
\description{
For the purposes of identifying text fields with no input in the Chicago
data, this function checks to see if a value is \code{NULL}, and more importantly
if it is simply a string of length zero or whitespace only.
}
