% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{str_split1}
\alias{str_split1}
\title{Separate a single string into a single character vector}
\usage{
str_split1(str, pattern)
}
\arguments{
\item{str}{A single character value}

\item{pattern}{A regular expression on which to split the string}
}
\value{
A character vector containing the split values
}
\description{
\code{str_split1()} allows the user to avoid problems associated with using
\code{stringr::str_split()}, which always returns a list, on a single character
value. Use the same was as you would \code{stringr::str_split()} but skip the need
to extract the first element from the result.
}
\examples{
> str_split1('Hello world')
[1] "Hello" "world"
}
