% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{subset_chifed_vars}
\alias{subset_chifed_vars}
\title{Subset a wide-format Chicago Fed data extract by variable name pattern}
\usage{
subset_chifed_vars(df_wide, pattern, negate = FALSE)
}
\arguments{
\item{df_wide}{A data frame extracted from a Chicago Fed data set}

\item{pattern}{A pattern to match variable names on}

\item{negate}{(default \code{FALSE}) Negate the match given by \code{pattern}?}
}
\value{
A \code{tibble} with the requested variables along with the necessary
identifying columns
}
\description{
Because there are so many variables in the Chicago Fed data set, there are
millions of rows in the resulting data set after pivoting into long form.
Splitting the data up by a variable name pattern allows us to store variables
with similar names in their own table, reducing the length of table scans
during queries and speeding up return of results.
}
\details{
For matching \emph{ALL} members of a variable group like "RSSD", "RCON", "RCFD",
etc., it is best to match on the first two letters because of how the data
is disaggregated, with each layer of reporting being given a separate 6-digit
suffix on a  two-letter prefix instead of the standard 4-letter, 4-digit
variable code.
}
\examples{
df_rcon <- subset_chifed_vars(df_wide, '^RC.{6}') ## Subset to RCON variables
df_text <- subset_chifed_vars(df_wide, '^TE.{6}') ## Subset to TEXT variables
}
