% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{update_cr_config_settings}
\alias{update_cr_config_settings}
\title{Update the package configuration with new parameters}
\usage{
update_cr_config_settings(..., cfg_file = build_config_filename())
}
\arguments{
\item{cfg_file}{The path to a YAML configuration file.}
}
\value{
The updated package settings
}
\description{
Update the package configuration with new parameters
}
