% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ffiec.R
\name{write_ffiec_varcodes}
\alias{write_ffiec_varcodes}
\title{Add discovered FFIEC variable names to an index table}
\usage{
write_ffiec_varcodes(db_connector, var_codes)
}
\arguments{
\item{db_connector}{}

\item{var_codes}{}
}
\description{
\code{write_ffiec_varcodes()} takes a list of variable codes and adds each of them
to a database table containing only a pairing between an integer ID and the
variable code. If the variable code is already in the table, it is skipped.
If it is not, it is automatically assigned an integer ID value in the order
in which it is added.
}
\details{
Given the numerous variables in the source data, this library pivots the data
into long form before writing to the database. This offers the benefit of
allowing each table in the database to have a predictable description and
prevent issues with exceeding the native maximum column support of many
database engines. However, it
}
