% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_metadata.R
\name{write_varcodes}
\alias{write_varcodes}
\title{Add variable names to an index table}
\usage{
write_varcodes(db_connector, var_codes)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{var_codes}{}
}
\description{
\code{write_varcodes()} takes a list of variable codes and adds each of them
to a database table containing only a pairing between an integer ID and the
variable code. If the variable code is already in the table, it is skipped.
If it is not, it is automatically assigned an integer ID value in the order
in which it is added.
}
\details{
Given the numerous variables in the source data, this library pivots the data
into long form before writing to the database. This offers the benefit of
allowing each table in the database to have a predictable description and
prevent issues with exceeding the native maximum column support of many
database engines.

However, pivoting the data creates a text column with a copy of the variable
code being stored for every one of the millions of observations in the data
set, significantly expanding the storage requirement of the database, and
potentially making queries run noticeably slower on standard spinning-platter
hard drives due to the increased scanning time.

This function aims to alleviate that problem by assigning each variable code
an integer ID value so that the storage requirements are significantly
smaller.
}
