% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ffiec_db.R
\name{zipped_schedules_not_in_db}
\alias{zipped_schedules_not_in_db}
\title{How much of this ZIP file has been extracted to the database?}
\usage{
zipped_schedules_not_in_db(db_connector, zip_file)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}
}
\value{
A \code{tibble} containing a record for each period, schedule, and part
code already extracted to the database, as determined by a query of the
\code{SUMMARY} table.
}
\description{
\code{compare_zip_to_db()} queries the \code{SUMMARY} table of the database, given by
the connector function \code{db_connector()}, to determine which schedules for
a particular period's FFIEC ZIP file have already been extracted to that
database, so as to allow \code{extract_all_ffiec_zips()} to begin where it left
off in its previous run.
}
