% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_chifed.R
\name{write_chifed_observations}
\alias{write_chifed_observations}
\title{Write a dataframe of Chicago Fed observations to the database}
\usage{
write_chifed_observations(df_wide, ymd_chr, suffix)
}
\arguments{
\item{db_connector}{A \code{function} created by one of the \verb{db_connector_*()}
functions found in this package. It should be passed without the \verb{()}}

\item{df_long}{A dataframe of observations in long form}
}
\description{
Write a dataframe of Chicago Fed observations to the database
}
